refactor(cli): move pkg/cli to go-clix/cli
Moves the pkg/cli package out of tree to https://github.com/go-clix/cli
