+++ cmd/tk/export.go
@@ -12,7 +12,8 @@ import (
 
 	"text/template"
 
-	"github.com/grafana/tanka/pkg/cli"
+	"github.com/go-clix/cli"
+
 	"github.com/grafana/tanka/pkg/tanka"
 )
