+++ cmd/tk/init.go
@@ -8,7 +8,8 @@ import (
 	"os"
 	"os/exec"
 
-	"github.com/grafana/tanka/pkg/cli"
+	"github.com/go-clix/cli"
+
 	"github.com/grafana/tanka/pkg/spec/v1alpha1"
 )
