+++ pkg/term/colordiff_test.go
@@ -1,4 +1,4 @@
-package cli
+package term
 
 import (
 	"io/ioutil"
